% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addon_import.R
\name{analyze_addon_import}
\alias{analyze_addon_import}
\title{Analyze Imported Addon Data}
\usage{
analyze_addon_import(data)
}
\arguments{
\item{data}{Data imported via import_addon_export()}
}
\value{
A list with analysis results
}
\description{
Run common analyses on data imported from ankiR Stats addon.
}
\examples{
\dontrun{
data <- import_addon_export("ankir_export.json")
analysis <- analyze_addon_import(data)
analysis$retention_trend
}
}
