% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{anki_base_path}
\alias{anki_base_path}
\title{Get Anki base path}
\usage{
anki_base_path()
}
\value{
Character string path to Anki2 directory
}
\description{
Returns the default Anki2 directory for the current platform.
}
\examples{
\dontrun{
anki_base_path()
}
}
