% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{anki_db_path}
\alias{anki_db_path}
\title{Get path to Anki database}
\usage{
anki_db_path(profile = NULL, base_path = NULL)
}
\arguments{
\item{profile}{Profile name (first profile if NULL)}

\item{base_path}{Path to Anki2 directory (auto-detected if NULL)}
}
\value{
Character string path to collection.anki2
}
\description{
Get path to Anki database
}
\examples{
\dontrun{
anki_db_path()
}
}
