% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_formats.R
\name{anki_export_importable}
\alias{anki_export_importable}
\title{Export to Anki-importable format}
\usage{
anki_export_importable(data, file, tags = NULL)
}
\arguments{
\item{data}{Data frame with front and back columns}

\item{file}{Output file path}

\item{tags}{Optional tags}
}
\value{
Invisibly returns number of cards
}
\description{
Export to Anki-importable format
}
\examples{
\dontrun{
cards <- data.frame(front = c("Q1", "Q2"), back = c("A1", "A2"))
anki_export_importable(cards, "new_cards.txt")
}
}
