% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media.R
\name{anki_media_path}
\alias{anki_media_path}
\title{Get media folder path}
\usage{
anki_media_path(path = NULL, profile = NULL)
}
\arguments{
\item{path}{Path to collection.anki2 (auto-detected if NULL)}

\item{profile}{Profile name (first profile if NULL)}
}
\value{
Character string path to collection.media folder
}
\description{
Get media folder path
}
\examples{
\dontrun{
anki_media_path()
}
}
