% Generated by roxygen2: do not edit by hand
\name{anki_plot_forgetting_curve}
\alias{anki_plot_forgetting_curve}
\title{Plot forgetting curve}
\usage{
anki_plot_forgetting_curve(curve_data)
}
\arguments{
\item{curve_data}{Output from anki_fit_forgetting_curve()}
}
\value{
A ggplot2 object
}
\description{
Visualize fitted forgetting curve compared to FSRS default.
}
\examples{
\dontrun{
curve <- anki_fit_forgetting_curve()
anki_plot_forgetting_curve(curve)
}
}
