% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{anki_suspended}
\alias{anki_suspended}
\title{Get suspended cards}
\usage{
anki_suspended(path = NULL, profile = NULL, include_notes = FALSE)
}
\arguments{
\item{path}{Path to collection.anki2 (auto-detected if NULL)}

\item{profile}{Profile name (first profile if NULL)}

\item{include_notes}{If TRUE, join with note data}
}
\value{
A tibble of suspended cards
}
\description{
Get suspended cards
}
\examples{
\dontrun{
suspended <- anki_suspended()
}
}
