% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{anki_timestamp_to_date}
\alias{anki_timestamp_to_date}
\title{Convert Anki timestamp to date}
\usage{
anki_timestamp_to_date(x)
}
\arguments{
\item{x}{Numeric timestamp in milliseconds since Unix epoch (1970-01-01)}
}
\value{
Date object
}
\description{
Converts Anki's millisecond-epoch timestamps to Date objects.
}
\examples{
# Convert a typical Anki timestamp
anki_timestamp_to_date(1368291917470)
}
