% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{anki_timestamp_to_datetime}
\alias{anki_timestamp_to_datetime}
\title{Convert Anki timestamp to datetime}
\usage{
anki_timestamp_to_datetime(x)
}
\arguments{
\item{x}{Numeric timestamp in milliseconds since epoch}
}
\value{
POSIXct datetime object
}
\description{
Convert Anki timestamp to datetime
}
\examples{
anki_timestamp_to_datetime(1368291917470)
}
