% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{anki_ts_plot}
\alias{anki_ts_plot}
\title{Plot time series with trend line}
\usage{
anki_ts_plot(ts_data, y_col, title = NULL)
}
\arguments{
\item{ts_data}{A tibble from any anki_ts_* function}

\item{y_col}{Column name to plot on y-axis}

\item{title}{Plot title}
}
\value{
A ggplot object
}
\description{
Plot time series with trend line
}
\examples{
\dontrun{
ts <- anki_ts_intervals()
anki_ts_plot(ts, "median_ivl", "Median Interval Over Time")
}
}
