% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_to_anki_timestamp}
\alias{date_to_anki_timestamp}
\title{Convert date to Anki timestamp}
\usage{
date_to_anki_timestamp(x)
}
\arguments{
\item{x}{Date or POSIXct object}
}
\value{
Numeric timestamp in milliseconds since Unix epoch
}
\description{
Converts a Date or POSIXct to Anki's millisecond-epoch format.
}
\examples{
date_to_anki_timestamp(Sys.Date())
}
