% Generated by roxygen2: do not edit by hand
\name{fsrs_from_csv}
\alias{fsrs_from_csv}
\title{Read FSRS Data from CSV}
\usage{
fsrs_from_csv(file, date_col = "date", rating_col = "rating",
              card_col = "card_id", date_format = "\%Y-\%m-\%d")
}
\arguments{
\item{file}{Path to CSV file with FSRS review data}
\item{date_col}{Name of date column}
\item{rating_col}{Name of rating column}
\item{card_col}{Name of card ID column}
\item{date_format}{Date format string}
}
\value{
A tibble with parsed FSRS review data
}
\description{
Read review data from a CSV file exported by FSRS optimizer or similar tools.
}
\examples{
\dontrun{
data <- fsrs_from_csv("revlog.csv")
}
}
