% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{fsrs_retrievability}
\alias{fsrs_retrievability}
\title{Calculate current retrievability for FSRS cards}
\usage{
fsrs_retrievability(stability, days_since_review, decay = 0.256)
}
\arguments{
\item{stability}{Stability in days}

\item{days_since_review}{Days since last review}

\item{decay}{Decay parameter (default 0.256)}
}
\value{
Retrievability (0-1)
}
\description{
Calculate current retrievability for FSRS cards
}
