% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{plot.anki_decomposition}
\alias{plot.anki_decomposition}
\title{Plot time series decomposition}
\usage{
\method{plot}{anki_decomposition}(x, ...)
}
\arguments{
\item{x}{An anki_decomposition object}

\item{...}{Additional arguments (unused)}
}
\value{
A ggplot object (if ggplot2 available) or base R plots
}
\description{
Plot time series decomposition
}
