% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamification.R
\name{print.anki_gamification}
\alias{print.anki_gamification}
\title{Print Gamification Stats}
\usage{
\method{print}{anki_gamification}(x, ...)
}
\arguments{
\item{x}{Gamification stats object}

\item{...}{Additional arguments}
}
\description{
Print Gamification Stats
}
