% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_formats.R
\name{fsrs_from_csv}
\alias{fsrs_from_csv}
\title{Import review data from CSV for analysis}
\usage{
fsrs_from_csv(
  file,
  date_col = "date",
  rating_col = "rating",
  card_col = "card_id",
  date_format = "\%Y-\%m-\%d"
)
}
\arguments{
\item{file}{Path to CSV file}

\item{date_col}{Name of date column}

\item{rating_col}{Name of rating column}

\item{card_col}{Name of card ID column}

\item{date_format}{Date format string}
}
\value{
A tibble with standardized review data
}
\description{
Import review data from CSV for analysis
}
\examples{
\dontrun{
reviews <- fsrs_from_csv("reviews.csv", date_col = "date", rating_col = "grade")
}
}
