% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_enhanced.R
\name{print.anki_mc_forecast}
\alias{print.anki_mc_forecast}
\title{Print Monte Carlo Forecast}
\usage{
\method{print}{anki_mc_forecast}(x, ...)
}
\arguments{
\item{x}{An anki_mc_forecast object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Monte Carlo Forecast
}
